흩어진 중복된 함수들 모아서 postFormatFunctions.ts에 작성
