+++ apps/client/src/util/format.ts
@@ -44,30 +44,3 @@ export const formatMinorJobToLabel = (input: string) => {
   }
 };
 
-export const formatDate = (dateString: string): string => {
-  const date = new Date(dateString);
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0'); // 월은 0부터 시작하므로 +1
-  const day = String(date.getDate()).padStart(2, '0');
-  return `${year}.${month}.${day}`;
-};
-
-export const calculateDDay = (dateString: string): number => {
-  // 현재 날짜
-  // 밀리초 단위 차이 계산
-  return Math.ceil(
-    (new Date(dateString).getTime() - new Date().getTime()) /
-      (1000 * 60 * 60 * 24),
-  ); // 일 단위로 변환
-};
-
-export const getEmploymentStatus = (employmentEndDate: string): string => {
-  if (employmentEndDate === '') return '상시';
-
-  const daysLeft = Math.ceil(
-    (new Date(employmentEndDate).getTime() - new Date().getTime()) /
-    (1000 * 60 * 60 * 24),
-  );
-
-  return daysLeft >= 0 ? `D-${daysLeft}` : '마감';
-};
